backend/requirements_sqlite.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
# pymysql==1.1.0  # 注释掉MySQL驱动
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
# redis==5.0.1  # 注释掉Redis
python-dotenv==1.0.0
